Module 4 - P2HW2
design a program that does the following:

Asks the user to enter grades for following tests. Each grades is to be requested in a separate statement (20 points)
Module 1
Module 2
Module 3
Module 4
Module 5
Module 6
The program should store the grades entered in a list. Give the list created a descriptive name (20 points)
Display the following: (10 points each)
The lowest grade in the list
The highest grade in the list
Sum of grades
The grades' average